elgamal: refactor ballot marshal/unmarshal
Better support to serialize/deserialize on JSON and CBOR for the Ballot an Ciphertext types.
